Revert "we should devide on int but not floor, or the text will be out of shape" I was completely wrong, we should divide on the floor or the text won't keep it's original size
This reverts commit 73f1b6a320f90cfec9de8878ec06fa351966beaf.

+++ gui/label.go
@@ -100,12 +100,12 @@ func (l *Label) SetText(text string) {
 	}
 
 	// Update label panel dimensions
-	width, height := textImage.Rect.Dx(), textImage.Rect.Dy()
+	width, height := float32(textImage.Rect.Dx()), float32(textImage.Rect.Dy())
 	if runtime.GOOS == "darwin" {
 		// since we enlarged the font texture for higher quality, we have to scale it back to it's original point size
 		width, height = width / 2, height / 2
 	}
-	l.Panel.SetContentSize(float32(width), float32(height))
+	l.Panel.SetContentSize(width, height)
 }
 
 // Text returns the label text.
